change vec4 in_joint to uvec4

+++ TP/shaders/basic.vert
@@ -7,7 +7,7 @@ layout(location = 1) in vec3 in_normal;
 layout(location = 2) in vec2 in_uv;
 layout(location = 3) in vec4 in_tangent_bitangent_sign;
 layout(location = 4) in vec3 in_color;
-layout(location = 5) in vec4 in_joint;
+layout(location = 5) in uvec4 in_joint;
 layout(location = 6) in vec4 in_weight;
 
 layout(location = 0) out vec3 out_normal;
